+++ NEWS.md
@@ -3,6 +3,7 @@
 
 | Date                  | News Item                                                     |
 | --------------------- | ------------------------------------------------------------- |
+| 2nd September 2025    | Release of [cstring 4.0.12](https://github.com/synesissoftware/cstring/releases/tag/4.0.12) |
 | 23rd February 2025    | Release of cstring 4.0.11                                     |
 | 25th October 2024     | Release of cstring 4.0.10                                     |
 | Available from [**cstring** project on GitHub](https://synesissoftware.com/cstring):  |